+++ src/index.ts
@@ -2,5 +2,6 @@ export { Client, ClientError } from "./client";
 export type { ClientOptions } from "./client";
 export { Server } from "./server";
 export type { ServerHandler, ServerOptions } from "./server";
+export { ModelProvider, Session } from "./server_new";
 export { sseEventsToMessages, resolvePendingToolUse } from "./utils";
 export type * from "./types";